Agentic_Journey/requirements.txt
streamlit
chromadb
sentence-transformers
google-genai
youtube-transcript-api